Separate player component.
