+++ src/app/pages/apps/pages/audio/components/audioevents/audioevents.component.html
@@ -5,10 +5,5 @@
 </ng-template>
 
 <div *ngIf="events.length; else loading">
-    <div *ngIf="downloadProgress | async as value">
-        <ngb-progressbar type="success" [value]="value"></ngb-progressbar>
-    </div>
-    <button type="button" class='btn btn-primary' (click)="renderer.togglePlayback()">Toggle playback</button>
-    <div #elapsed></div>
     <dng-timeline *ngFor="let event of events" [timelineLength]="timelineLength" [event]="event"></dng-timeline>
 </div>
