+++ src/app/pages/apps/pages/audio/components/index.ts
@@ -1,2 +1,3 @@
 export * from './timeline-entry';
 export * from './audioevents';
+export * from './player';
