+++ src/app/pages/apps/pages/audio/components/player/index.ts
@@ -1,0 +1,1 @@
+export * from './player.component';
