+++ src/app/utils/base/base.observable.spec.ts
@@ -1,7 +1,7 @@
 import { forkJoin } from 'rxjs';
 
-import { AppObservableArray, AppObservableArrayModes,
-    AppObservableObject, AppObservablePersistentObject } from './base.observable';
+import { AppObservableArray, AppObservableArrayModes, AppObservableObject,
+    AppObservablePersistentObject, NumberValueObserver } from './base.observable';
 
 
 class TestObservable extends AppObservableArray<string> {}
@@ -82,6 +82,21 @@ describe('TestObservable', () => {
 });
 
 
+describe('NumberValueObserver', () => {
+
+    it('subscription works', () => {
+        const testObj = new NumberValueObserver();
+
+        let value = 0;
+        testObj.observable.subscribe(_value => value = _value);
+
+        // Adding new value should succeed
+        testObj.setObject(3);
+        expect(value).toEqual(3);
+    });
+});
+
+
 class SubjectObservable extends AppObservableObject<string> {}
 
 describe('SubjectObservable', () => {
@@ -94,6 +109,24 @@ describe('SubjectObservable', () => {
 
         // Adding new item should succeed
         testObj.setObject('bar');
+        expect(state).toEqual('bar');
+    });
+
+    it('object subscription using asPipe method works', () => {
+        const testObj = new SubjectObservable();
+
+        let state = null;
+        testObj.asPipe().subscribe(_state => state = _state);
+
+        // Adding new item should succeed
+        testObj.setObject('bar');
+        expect(state).toEqual('bar');
+
+        // When observer is closed from emitting new values
+        testObj.closeSubject();
+
+        // New values are no longer received
+        testObj.setObject('foo');
         expect(state).toEqual('bar');
     });
 });
